PathPlanner: Planner now updates q map based off of the q learning state the root initializes the planner in

+++ ConsoleApplication1/QMap.cpp
@@ -6,12 +6,12 @@ QMap::QMap()
 {
 }
 
-int QMap::GetQ(Position i_position, Action i_action) const
+double QMap::GetQ(Position i_position, Action i_action) const
 {
 	return GetQ(R(i_position, i_action));
 }
 
-int QMap::GetQ(R i_r) const
+double QMap::GetQ(R i_r) const
 {
 	auto it = m_qMap.find(i_r);
 	if (it != m_qMap.end())
@@ -20,7 +20,7 @@ int QMap::GetQ(R i_r) const
 	return -1;
 }
 
-void QMap::SetQ(R i_r, int i_reward)
+void QMap::SetQ(R i_r, double i_reward)
 {
 	m_qMap[i_r] = i_reward;
 }
